+++ src/main/java/hoggaster/oanda/responses/OandaPrice.java
@@ -3,21 +3,22 @@ package hoggaster.oanda.responses;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 public class OandaPrice {
 
     public final String instrument;
-    public final Double bid;
-    public final Double ask;
+    public final BigDecimal bid;
+    public final BigDecimal ask;
     public final Date time;
     public final String status;
 
     @JsonCreator
     public OandaPrice(
             @JsonProperty(value = "currencyPair") String instrument,
-            @JsonProperty(value = "bid") Double bid,
-            @JsonProperty(value = "ask") Double ask,
+            @JsonProperty(value = "bid") BigDecimal bid,
+            @JsonProperty(value = "ask") BigDecimal ask,
             @JsonProperty(value = "time") Date time,
             @JsonProperty(value = "status", required = false) String status) {
         this.instrument = instrument;
